requirements.txt
anyio==3.7.1
bcrypt==4.0.1
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.105.0
FastAPI-SQLAlchemy==0.2.1
fastapi-utils==0.2.1
greenlet==3.0.2
h11==0.14.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
psycopg2==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.13
pytest==7.4.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
text-unidecode==1.3
typing_extensions==4.9.0
uvicorn==0.24.0.post1
